+++ CHANGELOG.md
@@ -13,6 +13,8 @@ Given a version number MAJOR.MINOR.PATCH, increment:
 
 
 ## [Unreleased]
+
+## [2.14.0] - 2025-07-30
 ### Added
 - display_description parameter to Transfer resource